(AuthToken) change Token only on expire

+++ include/Webservices/AuthToken.php
@@ -11,24 +11,42 @@
 function vtws_getchallenge($username){
 	global $adb;
 
+	$authToken = uniqid();
+	$servertime = time();
+	$expireTime = time()+(60*5);
+
 	$user = new Users();
 	$userid = $user->retrieve_user_id($username);
 	if (empty($userid)) {
 		throw new WebServiceException(WebServiceErrorCode::$AUTHREQUIRED,'Given user cannot be found');
 	}
 
-	$authToken = uniqid();
+	$get_token = $adb->pquery("SELECT * FROM vtiger_ws_userauthtoken WHERE userid = ?", array($userid) );
 
-	$servertime = time();
-	$expireTime = time()+(60*5);
+	if($adb->num_rows($get_token) == 1) {
+		$user_data = $adb->fetchByAssoc($get_token, 0);
+		$expired_time_unix = $user_data['expiretime'];
+		$expiretime = new DateTime("@$expired_time_unix"); 
 
-	$sql = "delete from vtiger_ws_userauthtoken where userid=?";
-	$adb->pquery($sql,array($userid));
+		$now = new DateTime();
+		$diff = $expiretime->diff($now);
 
-	$sql = "insert into vtiger_ws_userauthtoken(userid,token,expireTime) values (?,?,?)";
-	$adb->pquery($sql,array($userid,$authToken,$expireTime));
+		if($diff->invert == 0) {
+
+			$sql = "delete from vtiger_ws_userauthtoken where userid=?";
+			$adb->pquery($sql,array($userid));
+
+			$sql = "insert into vtiger_ws_userauthtoken(userid,token,expireTime) values (?,?,?)";
+			$adb->pquery($sql,array($userid,$authToken,$expireTime));
+
+		} else {
+			$authToken = $user_data['token'];
+			$expireTime = $expired_time_unix;
+		}
+	} else {
+		$sql = "insert into vtiger_ws_userauthtoken(userid,token,expireTime) values (?,?,?)";
+		$adb->pquery($sql,array($userid,$authToken,$expireTime));	
+	}
 
 	return array("token"=>$authToken,"serverTime"=>$servertime,"expireTime"=>$expireTime);
 }
-
-?>
